move ModelConfigParams out of openai
