+++ crates/goose/src/providers/githubcopilot.rs
@@ -9,8 +9,9 @@ use async_trait::async_trait;
 use axum::http;
 use chrono::{DateTime, Utc};
 use goose_providers::errors::ProviderError;
-use goose_providers::formats::openai::{is_openai_responses_model, ModelConfigParams};
+use goose_providers::formats::openai::is_openai_responses_model;
 use goose_providers::images::ImageFormat;
+use goose_providers::models::ModelConfigParams;
 use reqwest::{Client, Response};
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
