+++ crates/goose/src/providers/nanogpt.rs
@@ -9,8 +9,9 @@ use anyhow::Result;
 use async_trait::async_trait;
 use futures::future::BoxFuture;
 use goose_providers::errors::ProviderError;
-use goose_providers::formats::openai::{create_request, ModelConfigParams};
+use goose_providers::formats::openai::create_request;
 use goose_providers::images::ImageFormat;
+use goose_providers::models::ModelConfigParams;
 use rmcp::model::Tool;
 
 pub const NANOGPT_PROVIDER_NAME: &str = "nano-gpt";
