+++ crates/goose/src/providers/ollama.rs
@@ -17,8 +17,8 @@ use async_trait::async_trait;
 use futures::future::BoxFuture;
 use futures::TryStreamExt;
 use goose_providers::errors::ProviderError;
-use goose_providers::formats::openai::ModelConfigParams;
 use goose_providers::images::ImageFormat;
+use goose_providers::models::ModelConfigParams;
 use reqwest::Response;
 use rmcp::model::Tool;
 use serde_json::{json, Value};
