+++ crates/goose/src/providers/openrouter.rs
@@ -2,6 +2,7 @@ use anyhow::Result;
 use async_trait::async_trait;
 use futures::future::BoxFuture;
 use goose_providers::images::ImageFormat;
+use goose_providers::models::ModelConfigParams;
 use serde_json::{json, Value};
 
 use super::api_client::{ApiClient, AuthMethod};
@@ -13,7 +14,7 @@ use crate::conversation::message::Message;
 use crate::model::ModelConfig;
 use crate::providers::formats::openrouter as openrouter_format;
 use goose_providers::errors::ProviderError;
-use goose_providers::formats::openai::{create_request, ModelConfigParams};
+use goose_providers::formats::openai::create_request;
 use rmcp::model::Tool;
 
 pub const OPENROUTER_PROVIDER_NAME: &str = "openrouter";
