+++ crates/goose/src/providers/toolshim.rs
@@ -42,8 +42,8 @@ use anyhow::Result;
 use futures::StreamExt;
 use goose_providers::errors::ProviderError;
 use goose_providers::formats::openai::create_request;
-use goose_providers::formats::openai::ModelConfigParams;
 use goose_providers::images::ImageFormat;
+use goose_providers::models::ModelConfigParams;
 use reqwest::Client;
 use rmcp::model::{object, CallToolRequestParams, RawContent, Tool};
 use serde_json::{json, Value};
